removed depricated sdss file link generation

+++ aspectui.py
@@ -42,11 +42,6 @@ zoomed_som_directory = '/'.join((output_directory, "icons"))
 if not os.path.exists(zoomed_som_directory):
     os.makedirs(zoomed_som_directory)
 
-#sdsslink file
-sdsslinks_directory = '/'.join((output_directory, "sdsslinks"))
-if not os.path.exists(sdsslinks_directory):
-    os.makedirs(sdsslinks_directory)
-
 #mapping from som_x, som_y to mjd, plateid, fiberid 
 idmapping_directory = '/'.join((output_directory, "idmapping"))
 if not os.path.exists(idmapping_directory):
@@ -103,25 +98,21 @@ for row in tr:
                 plateid = -1
                 fiberid = -1                
             
-            #write sdsslink file
-            #deprecated
-            sdsslink_file_path = ''.join((sdsslinks_directory, '/', str(som_x), '-', str(som_y), '.link'))
-            sdsslink_output_string = ''.join(('MPF: ', str(mjd), '-', str(plateid), '-', str(fiberid), ' <a href="', link, '" target="_blank">Explore</a>'))
-            with open(sdsslink_file_path, 'w') as sdsslink_file:
-                sdsslink_file.write(sdsslink_output_string)
-            
             #spectra meta data
             spec_meta_data_file_path = ''.join((spec_meta_data_directory, '/', str(som_x), '-', str(som_y), '.json'))
-            with open(spec_meta_data_file_path, 'w') as spec_meta_data_file:
-                json.dump({"mjd": int(mjd), "plateid":int(plateid), "fiberid": int(fiberid), "sdsslink": link, "som_x": int(som_x), "som_y": int(som_y)}, spec_meta_data_file)
+            if not os.path.exists(spec_meta_data_file_path):
+                with open(spec_meta_data_file_path, 'w') as spec_meta_data_file:
+                    json.dump({"mjd": int(mjd), "plateid":int(plateid), "fiberid": int(fiberid), "sdsslink": link, "som_x": int(som_x), "som_y": int(som_y)}, spec_meta_data_file)
             
             #write idmapping som_x, som_y -> mjd, plateid, fiberid and vice versa
             idmapping_file_path = ''.join((idmapping_directory, '/', str(som_x), '-', str(som_y), '.json'))
-            with open(idmapping_file_path, 'w') as idmapping_file:
-                json.dump({"mjd": int(mjd), "plateid":int(plateid), "fiberid": int(fiberid) }, idmapping_file)
+            if not os.path.exists(idmapping_file_path):
+                with open(idmapping_file_path, 'w') as idmapping_file:
+                    json.dump({"mjd": int(mjd), "plateid":int(plateid), "fiberid": int(fiberid) }, idmapping_file)
             idmapping_file_path = ''.join((idmapping_directory, '/', str(mjd), '-', str(plateid), '-', str(fiberid), '.json'))
-            with open(idmapping_file_path, 'w') as idmapping_file:
-                json.dump({"som_x": int(som_x), "som_y":int(som_y)}, idmapping_file)            
+            if not os.path.exists(idmapping_file_path):
+                with open(idmapping_file_path, 'w') as idmapping_file:
+                    json.dump({"som_x": int(som_x), "som_y":int(som_y)}, idmapping_file)           
             
             
             #make empty.png in highest zoom level if it not exists
